+++ packages/fraxel/src/nodes/lib/types.ts
@@ -8,6 +8,7 @@ import type { RayCast, RayCastOptions } from '../node2d/ray-cast.js'
 import type { PrimaryNode } from './enum.js'
 import type { Clickable, ClickableOptions } from '../node2d/clickable.js'
 import type { Draggable, DraggableOptions } from '../node2d/draggable.js'
+import type { DropArea, DropAreaOptions } from '../node2d/drop-area.js'
 import type { Timer, TimerOptions } from '../timer.js'
 import type { Geometry, GeometryOptions } from '../node2d/geometry.js'
 import type { Text, TextOptions } from '../node2d/text.js'
@@ -42,6 +43,7 @@ export interface NodeClasses {
   [PrimaryNode.RayCast]: typeof RayCast
   [PrimaryNode.Clickable]: typeof Clickable
   [PrimaryNode.Draggable]: typeof Draggable
+  [PrimaryNode.DropArea]: typeof DropArea
   [PrimaryNode.Timer]: typeof Timer
   [PrimaryNode.Geometry]: typeof Geometry
   [PrimaryNode.Text]: typeof Text
@@ -74,6 +76,7 @@ export interface NodesOptions {
   [PrimaryNode.RayCast]: RayCastOptions
   [PrimaryNode.Clickable]: ClickableOptions
   [PrimaryNode.Draggable]: DraggableOptions
+  [PrimaryNode.DropArea]: DropAreaOptions
   [PrimaryNode.Timer]: TimerOptions
   [PrimaryNode.Geometry]: GeometryOptions
   [PrimaryNode.Text]: TextOptions